cryptohome: Protect TpmPersistentState data with lock
This protects data members in TpmPersistentState with a lock to make
it safe to use from multiple threads. Reality is that this object is
accessed by multiple threads, and there isn't a clear path towards
restricting TPM state tracking to only a single thread as a potential
alternative option.

BUG=chromium:908393
TEST=tsan no longer reports data races on TpmPeristentState::IsTpmReady()

Change-Id: I746afc30b184d9463cd3232c0f0341dffb448205
Reviewed-on: https://chromium-review.googlesource.com/1350969
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>