shill: fixup Nl80211Frame parsing
The frame length handling in the Nl80211Frame() constructor was a little
imprecise. It checked for a minimum length that included the header and
16 bits of body, but some of these frames utilize fields that are past
the 16-bit boundary.

Let's refactor the struct to group the header and body fields
differently, and split the bounds-checking to
(a) initial header check and
(b) per-frame-type body length check

Doing this kills the general "kMinimumFrameByteCount", and we just
require that all frames have the 24-byte minimum header instead.

Along the way, I also
 * fix the endianness handling on frame_control (we need to convert
   fields larger than a byte from LE to native endianness)
 * avoid duplicating the frame_type extraction in ToString()

The latter point helps avoid duplicating the bounds checks too, since we
can just declare the "type" as invalid in the constructor.

Lastly, make the packed alignment explicit and add a couple of sanity
checks in the unit tests. We currently are at the mercy of the compiler
to keep our 16-bit / 8-bit fields matching the 802.11 frame format.

BUG=chromium:907847
TEST=unit tests; fuzzer targets; watch net.log on a device

Change-Id: I87f9a5e82d586f0f3fac87cf96a93acc30d84b3d
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1351120
Reviewed-by: Ben Chan <benchan@chromium.org>