authpolicy: Back up and restore user realm
If authpolicyd is restarted mid-session for some reason (manually or due
to a crash), and the user triggers a policy fetch, authpolicyd runs into
a CHECK in SambaInterface::FetchUserGpos() since the user realm is not
backed up and restored. This CL adds this as well as a regression test.

BUG=chromium:908772
TEST=On a Chromad device, log in, open console, type 'restart
     authpolicyd', open Chrome, go to chrome://policy, click "Reload
     policies", check /var/log/authpolicyd. It should NOT say
     "Check failed: !user_account_.realm.empty()".

Change-Id: Ic08f52d751b46983a2becb8407b319598a91f90f
Reviewed-on: https://chromium-review.googlesource.com/1352169
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Commit-Ready: Roman Sorokin <rsorokin@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Tested-by: Roman Sorokin <rsorokin@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>