chromeos-config: Support config partials for dump
For partial configs (e.g. public versus private), sometimes no identity
is specified in the private config portion.  We're just using this for
unique hashing, which the name works fine for also.

This is to support the breakup of public/private config dumps for coral.

BUG=b:119520306
TEST=FEATURES=test emerge-coral chromeos-config

Change-Id: If325c9669b92223a4e168a98d95e5aa98415da2f
Reviewed-on: https://chromium-review.googlesource.com/1352224
Commit-Ready: C Shapiro <shapiroc@chromium.org>
Tested-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>