chromeos-config: Fix issue with dptf-dv param
Originally, it didn't look like dptf-dv was used anywhere, but it ended
up embedded in a Chrome binary payload that does end up on the target.
See bug for context.

Adding it back to diff'ing support since this is a valid parameter.

BUG=b:120088670
TEST=local build/test on coral and diff testing
CQ-DEPEND=CL:1352372

Change-Id: I7d5a3a850e6548fee1128d506661eb6faf900198
Reviewed-on: https://chromium-review.googlesource.com/1352228
Commit-Ready: C Shapiro <shapiroc@chromium.org>
Tested-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Gregory Meinke <gmeinke@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>