hal_adapter: reprocessing effects call device ops with device ops thread
The HAL adapter spawns a thread to apply the reprocessing effects so
that it does not block other requests. If it also needs the HAL to
process the request, it is delegated back to the device ops thread so
that all device ops are called with the same thread.

BUG=b:118415134,b:118469770
TEST=Take portrait mode pictures on nocturne

Change-Id: I6d452d132e51bd71417262c8cdaf5f7f8a354074
Reviewed-on: https://chromium-review.googlesource.com/1351865
Commit-Ready: Hung-yu Wu <hywu@chromium.org>
Tested-by: Hung-yu Wu <hywu@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>