authpolicy: Cache GPOs if version did not change
Reuses cached GPO files if their version did not change. The daemon
already keeps GPO files somewhere in a tmpfs-mounted /tmp folder. Note
that this will NOT cache GPO files between logins (logging out restarts
authpolicyd and hence wipes the cache).

Wipes the cache every 75 hours. Note that the number is not a multiple
of 24 hours as you don't want users to wait longer for policy fetch
every N-th day in the morning when they log in for the first time.

Right now, the cache cannot be turned off, but there will be a device
policy in the future that controls the TTL and can be used to toggle the
cache.

BUG=chromium:908342
TEST=Log in on an Active Directory managed device:
     Open Crosh and enter 'authpolicy_debug 1'
     Open chrome://policy and reload policies a few times
     Open shell and enter 'grep "GPO Cache" /var/log/authpolicy.log'.
     Most recent logs should say "Using cached version..." and NOT
     "Downloading (not in cache)".

Change-Id: I1b8b3d85e0638ea5ad4080407f52a923c5e11c45
Reviewed-on: https://chromium-review.googlesource.com/1352421
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>