chaps: fix polymorphic SecureBlob usage
In preparation for breaking the dependence of SecureBlob
on Blob, create a temporary type SecureVector which
is essentially a Blob. This avoids compilation errors
in this CL.

Mechanical change to use SecureVector as the base
interface to pass to DBus bindings which rely on
template specialization. Additionally, fixup
function signatures to use either SecureBlobs
or SecureVectors.

BUG=chromium:728047
TEST=unittests.

Change-Id: I03487a86f6e7366525b65cbc5c83a0fda923afd1
Reviewed-on: https://chromium-review.googlesource.com/1357595
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>