authpolicy: whitelist 'authpolicyd-exec' user for UID change
CL:1055871 extends the chromiumos LSM to support configuring per-UID
policies in CrOS that restrict which other UIDs can be switched to by
processes spawned under the restricted UID. This policy also determines
the UIDs to which a user can set the saved set-UID. Configure a policy
to be applied to the authpolicyd user, so authpolicyd can only set its
saved set-UID to the 'authpolicyd-exec' user and can't perform any other
actions granted by CAP_SETUID. Check crbug.com/910316 for examples of
other system services we have restricted in CrOS with this LSM
functionality.

NOTE: with this change in, it would be fine to give authpolicyd full
CAP_SETUID rights since the LSM will ensure that the only thing it can
do with those rights is switch to the 'authpolicyd-exec' user. I haven't
done that refactoring in this CL.

CQ-DEPEND=CL:1355510
BUG=chromium:910316
TEST=enterprise_AuthPolicyDaemonServer.auth

Change-Id: I56a8c5c58068c61d5925a34cb1e77e88c876ee55
Reviewed-on: https://chromium-review.googlesource.com/1355511
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Micah Morton <mortonm@chromium.org>