shill: lock down CAP_SETUID for 'ipsec' and 'openvpn' users
The unprivileged parts of strongSwan and OpenVPN should not be able to
do anything with CAP_SETUID, which they will have inherited from their
parent programs that needed the capability in order to drop from 'shill'
user. We could patch these 3rd party programs to drop the CAP_SETUID
capability, but this is easier and just as affective.

BUG=chromium:649417
TEST=modified
/usr/share/cros/startup/process_management_policies/shill_whitelist.txt
and the security_ProcessManagementPolicy autotest to make sure 'ipsec'
and 'openvpn' can't setuid to other users.

Change-Id: Ibfdf6433bb4f114e1a0a135f45851283b58ddd9e
Reviewed-on: https://chromium-review.googlesource.com/1356101
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>