libbrillo: Add ResolveHostToIp function for shill's use
We are planning to add an HTTPS probe into our captive portal
detection scheme. curl can only do HTTPS requests to a custom IP
and port if it resolves an HTTPS URL to that IP and port. This
AppendHost function will use the CURLOPT_RESOLVE option to force
curl to do this DNS resolution by adding this mapping to the DNS
cache.

BUG=chromium:903645
TEST=`FEATURES=test emerge-caroline libbrillo`

Change-Id: I9b0c5d614beaf025483f907f5614abd82c61f5f3
Reviewed-on: https://chromium-review.googlesource.com/1356359
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>