vm_tools: maitred: Send SIGPWR to lxd on shutdown
lxd treats SIGPWR as a signal that the system is going to lose power and
that it should shut down all containers and exit.  Send SIGPWR to lxd
before sending SIGTERM to all other processes during shutdown.

This will not improve things much for the stretch container due to an
issue with systemd but it should make shutdown for all other types of
containers much faster.

BUG=chromium:906804
TEST=concierge_client --stop_all

Change-Id: I79d0f618c7c0dc24a16d9d5fc7e7259acaa71f4a
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1357643
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>