arc: Refactor arc-networkd for multi-network support.
This is an update to patch https://crrev.com/c/1359634 which was reverted because it broke cheets_CleanShutDown{_P} autotests. It is identical except for code removed in arc_ip_config.cc around lines 225-227 and in arc-network.conf around line 75. This first fix restores the behavior of leaving the interface up and relying on the container itself to clean it up. The second is just a missed merge issue that is effectively a no-op but should not be there.

BUG=b:77293260
TEST=flashed to nocturne, verified was able to download apps from Play store, plugged/unplugged eth0 dongle - verified default interface switching still works, verified device bridge added/removed as expected. Verified cheets_CleanShutDown{_P} on nocturne pass.

Change-Id: I5ebc71d2db394078e6d9616dbc8b36bd6c689298
Reviewed-on: https://chromium-review.googlesource.com/1401481
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>