biod: End EnrollSession/AuthSession when session fails
When the session fails due to an error,
CrosFpBiometricsManager::OnSessionFailed() is called, which calls
BiometricsManagerWrapper::OnSessionFailed(). That method was cleaning up
the associated dbus objects, but not the session itself.

If a new session was started in this state, biod would think that a
session was already running (specifically,
CrosFpBioMetricsManager::next_session_action_ would be non-null) and
refuse to start a new session. With this change biod will correctly
allow a new session to be started after a failure since the session has
been cleaned up.

This change also adds some additional logging of the error cases to make
it more obvious.

BUG=b:119131588
TEST=enroll finger and unlock
TEST=delete enrolled finger
TEST=enroll with error
TEST=enroll finger and unlock

Change-Id: Ic78d156362d054bcfa142c1fae97e5183845b1f6
Reviewed-on: https://chromium-review.googlesource.com/1366637
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>