trunks: Remove Verify() in tpm_unility
In this moment, we don't use TPM-based Verify in reality.
It's always sign-with-TPM -> verify-in-software.

To minize the dead code in our code base, remove Verify() in
tpm_unility.

If we need Verify() in the feature, just revert this CL
(and CL:1366669 if needed) and add full tests.

BUG=None
TEST=unittest
CQ-DEPEND=CL:1366669

Change-Id: I3e939c98e1cbac802f0fd818d993d63cfbcc9c14
Reviewed-on: https://chromium-review.googlesource.com/1366670
Commit-Ready: Meng-Huan Yu <menghuan@chromium.org>
Tested-by: Meng-Huan Yu <menghuan@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>