vm_tools: Fix window parenting for transient windows
This fixes a bug where we wouldn't always set the parent window
correctly for transient windows with X11. Sometimes the child was
realized before the parent. Sometimes the direct parent doesn't exist as
a wl_surface so we need to find an alternate.

BUG=chromium:910219
TEST=Verified that VLC has dialogs on top and they stay there

Change-Id: I098ee22e66dcab4e7a438872b41007867743d8b0
Reviewed-on: https://chromium-review.googlesource.com/1368368
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: David Reveman <reveman@chromium.org>