power: powerd_suspend: exit sniff mode on resume
The bluetooth HCI Mode Change event is disabled in the following CLs
  - http://crrev.com/c/704458
  - http://crrev.com/c/908249
when system is being suspended to prevent the controller from waking
up the system accidentally.

When the system is suspended for a while, a connected bluetooth
device may determine to enter the Sniff mode to save power and
sends a Sniff mode change event to the controller. Such a Mode
Change event is disabled so as not to wake up the system.

Such a Mode Change event is only kept in the controller. The
software side is not aware of which mode a bluetooth device is
currently in on resume. If a device is in the Sniff mode, the audio
replaying would become stuttering. To work around the issue, let
powerd send exit-sniff-mode commands to all ACL connections.

Note that if an audio device is connected through SCO connection,
there would be no stuttering. This is because, on system resume,
kernel would re-initialize the SCO connection and thus bring the
device into the active mode.

BUG=b:120603675
TEST=Perform the following steps.
1. Use either ethernet or wifi to connect the DUT to Internet.
2. Connect one or two bluetooth headphones/speakers to the DUT.
3. Open a youtube page to listen to music.
4. Suspend the DUT for 1 minute and then resume by the command
   $ powerd_dbus_suspend --wakeup_timeout=60
5. Observe that there is no audio/video stuttering.

Change-Id: I2d0bf8414221ffc1be30ccf68675bb13df4d6036
Reviewed-on: https://chromium-review.googlesource.com/1370113
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Sukumar Ghorai <sukumar.ghorai@intel.com>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>