chaps: Isolate RSA key generation code to GenerateRSAKeyPair()
Move all RSA key generation related code out from
GenerateKeyPair() in session_impl which is the implementation of
PKCS#11 C_GenerateKeyPair. Then we can put the general stuff in
GenerateKeyPair() and dispatch to the corresponse function
according to the mechanism.

The shared RSA logic is put in GenerateRSAKeyPair(),
software-backed generation is put in GenerateRSAKeyPairSoftware(),
and TPM-backed is put in GenerateRSAKeyPairTPM().

BUG=chromium:910633
TEST=passed chaps unit tests (expected no change of behavior)
     passed manually test of p11_replay

Change-Id: I489e18c05855eca565bcd50599ae8b8242e22660
Reviewed-on: https://chromium-review.googlesource.com/1373593
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Meng-Huan Yu <menghuan@chromium.org>
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>