vm_tools: Set container timezone when started
When a container has finished starting, read the system timezone
from /etc/timezone and set the timezone of that container to match.

Change tremplin's SetTimezone to take a list of containers for which
we want to set the timezone.

Modifies the seccomp policies to allow lstat/readlink so that we can
follow the symlink at /etc/localtime to an actual TZif file under
/usr/share/zoneinfo.

Modifies TzifParser to allow absolute and relative paths.

BUG=chromium:829934
TEST=start a container using 'vmc container' and check that 'date'
returns the proper time. Then, change system timezone and re-start
the container, once again checking that 'date' returns the new time.
CQ-DEPEND=CL:1454938

Change-Id: I498345c30e6f80ec0c880ded1136195609bcd59b
Reviewed-on: https://chromium-review.googlesource.com/1454428
Commit-Ready: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>