shill: manage lifetime of SupplicantProcessProxy callback
This proxy object maintains its lifetime with the WiFi object, but it
also posts tasks to the event loop using this Closure. This means
there's a possibility of queueing tasks to the event loop that will run
after we are destroyed.

Instead of an unmanaged Unretained(), give the callback a weak pointer,
so the run loop can decide whether or not to run the callback based on
the weak pointer's validity.

This should avoid some nullptr dereferences we've seen, where
OnSupplicantAppear() is called just after its WiFi object was destroyed.

BUG=chromium:698457
TEST=unit tests; manual test - add/remove Wifi devices (e.g., with 'iw')

Change-Id: Ibd1b14ec51bee8a322ef36a586dd625273cd9a9f
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1377576
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>