vm_tools: Add flag for x11_demo app window title.
This adds a flag to set the x11_demo app window title. This is
useful to identify x11_demo app window instances during tests. The tast
tests are asynchronous between the controlling machine and the DUT.
Using a different title between invocations will make sure that the
right window is measured without the need for the control program to
sleep for a fixed time.

BUG=chromium:839242
TEST=Manually run the updated app and verified the title flag.
Change-Id: Ia9486b065b692964f3b94040f5ae8b72fb54453f
Reviewed-on: https://chromium-review.googlesource.com/1377577
Commit-Ready: Tim Zheng <timzheng@chromium.org>
Tested-by: Tim Zheng <timzheng@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>