cryptohome: Improve RemoveKeyset
RemoveKeyset tries to go through expensive authenticate operation first
before removing the keyset. Most of the time the keyset is not present
on the device. This CL, checks first if the keyset is present on the
device. Once the presence is confirmed it can go through credentials
check.

BUG=chromium:912126
TEST=None

Change-Id: I510459dcb05083f1b5918dfdbf4fda99646e6490
Reviewed-on: https://chromium-review.googlesource.com/1377419
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>