vm_tools: seneschal: Remove empty directories after unshare
When a path such as MyFiles/a/b/c is shared, we create empty
dirs for MyFiles/a/b/c and then mount into this point.
When unsharing, we should remove these empty directories,
but we must take care to only remove directories which are
not also part of another mount such as MyFiles/a/b2/c2.

BUG=chromium:909833
TEST=Test sharing with FilesApp on device

Change-Id: Iab7130d6c8cc1313bcd8f118a097aa713dc9e759
Reviewed-on: https://chromium-review.googlesource.com/1381312
Commit-Ready: Joel Hockey <joelhockey@chromium.org>
Tested-by: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>