vm_tools: Added exec to app data
This changes garcon and cicerone to send the exec names of desktop files
are appropriately communicated via protos. Note that the exec name is parsed
so that it is the exec name and not just the raw exec field.

TEST=updated desktop_file_test to test for exec, mannually tested overall
system

BUG=chromium:849936

CQ-DEPEND=CL:1381311

Change-Id: I60422da5649962193154cd163431d8d22ef4a19f
Reviewed-on: https://chromium-review.googlesource.com/1381320
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Daniel Ng <danielng@google.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>