init: new usermode-helper
This filters all upcalls the kernel makes to userspace.  This way we
can catch & reject all attempts to subvert configs that spawn calls
as the root user.

For now we trigger crash reports when the program is unknown but we
don't block it.  This way we can get feedback from the field before
we start blocking it for real.

BUG=chromium:915974
TEST=unittests & precq pass
TEST=booting a system with everything (inc kernel) enabled works & no crashes in /var/spool/crash

Change-Id: I2563b3b992780a024d8a853b4e02b84993eabc0c
Reviewed-on: https://chromium-review.googlesource.com/1381653
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>