power: powerd_suspend: disable some mask events on suspension
Different Bluetooth devices and Bluetooth controllers of
different vendors may send different events to the host during
system suspension which would wake up the system spuriously. These
events observed include but not limited to Mode Change event, Sniff
Subrating event, Max Slots Change event, etc. Disabling the events
on suspension would help prevent the system from waking up spuriously.

Note:
- Intel Bluetooth controller connected to a Bluetooth audio device,
  e.g., a Bose QC35 speaker, may receive a Mode Change event after
  suspension.
- QCA Bluetooth controller connected to the above Bluetooth audio
  device may receive a Mode Change event and also a Sniff Subrating
  event after suspension.
- QCA Bluetooth controller connected to other Bluetooth audio
  devices may receive other events, e.g., Max Slots Change event,
  after suspension.
- QCA Bluetooth controller may emit a few more other events if
  scanning is turned on on suspension.
- QCA Bluetooth controller may emit Number of Completed Packets event
  and ACL Data RX when connected to Mi Sports Bluetooth Earphones.

We have to be careful when disabling mask values to avoid the system
from going out of sync with the controller on resume. Here we turn
off the following specific events that would not suffer from the
out-of-sync problem. Disabling these events should take care of most
spurious wakeup issue.
- Mode Change Event and Sniff Subrating Event: these states will be
  synced again after the host sends Exit Sniff mode command on resume.
- Max Slots Change Event: kernel does not care this event at all.

The Number of Completed Packets event and ACL Data RX when connected
to Mi Sports Bluetooth Earphones are not handled in this patch as
discarding them would result in the undesired out-of-sync problem
between the host and the controller. If many devices suffer from this
issue, it should be handled with another method which is beyond the
scope of this patch.

BUG=b:119700112
TEST=Perform the following steps.
1. Connect a Bluetooth audio device to the DUT.
2. Play music, e.g., through a youtube page.
3. Make sure that there is no discovering/scanning going on.
   Let the system suspend for 60 seconds with the commands
   $ date; powerd_dbus_suspend --wakeup_timeout=60; date
4. Verify that system will not auto resume in 60 seconds.
5. Also verify whether the audio plays smoothly after resume.

Use DUTs with controllers from different vendors.

Change-Id: I69c40fa6a29469cec33c79a6f30c90f1dc41091f
Reviewed-on: https://chromium-review.googlesource.com/1383763
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Dmitry Grinberg <dmitrygr@google.com>