cryptohome: Improve ResetLECredentials
ResetLECredentials tries to go through expensive authenticate operation
before looking for vault keyset to reset. Many times user has no
LECredentials keyset, so the authentication just wastes time.
This change tries to check the credentials only if at least one
LECredentials keyset was found.

Another optimization that can be done, is to check if the failed
attempts counter in the keyset is greater than zero. Otherwise it makes
no sense to reset, so we could avoid authentication in that case too.
That is planned to be done in a separate CL.

BUG=chromium:912126
TEST=None

Change-Id: I7a9f45bbdaec53d49ad65b000192fecb2cbb7d8b
Reviewed-on: https://chromium-review.googlesource.com/1384310
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>