init: Add wrapper library for crossystem
Adds C++ style wrapper library for crossystem that allows for testing
by swapping to a fake implementation. The library is based on a version
under login_manager but is not identical.

With this, we can have clobber_state.cc use dependency injection,
allowing for easier and more readable testing.

BUG=chromium:884520
TEST=code builds

Change-Id: I61b10fe367eb030ce525f2ac421e1b935a2b6f4f
Reviewed-on: https://chromium-review.googlesource.com/1387205
Commit-Ready: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>