vm_tools: enable gpu with vmc command option
This change allows a user to start a VM with gpu enabled using an option
in `vmc start`. The intention is to open up this feature to a broader
set of testers than just the ones that can be bothered to recompile
large pieces of crostini and deploy them to a dev mode Chrome OS device.
Using the GPU in crostini is not ready for everyday use and should be
considered pre-alpha. The option to enable GPU can be taken away at any
time.

BUG=chromium:837073
TEST=vmc start --enable-gpu termina

Change-Id: I9f58f4b593fd768febf75d253ba552974c030119
Reviewed-on: https://chromium-review.googlesource.com/1387842
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>