camera: usb: Try to use minimum resolution for multiple streams
We used native ratio for different resolution streams. But for all
streams with the same aspect ratio we don't need to choose native ratio.
We can use the largest resolution of the streams and scale it to other
streams.
For native ratio resolution, this CL will choose the minimum resolution
suitable for all resolutions of streams.

BUG=b:117628953
TEST=Check we use 1920x1080 not 2560x1920 for 1920x1080 and 1280x720
streams on DUT.
Check we use 1280x960 not 2592x1944 for 640x480 and 1280x720 streams.

Change-Id: I0dca303069a5d35971f096b2398bc39db338df1f
Reviewed-on: https://chromium-review.googlesource.com/1436649
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Hsu Wei-Cheng <mojahsu@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>
Reviewed-by: Heng-ruey Hsu <henryhsu@google.com>