shill: (retry) manage lifetime of SupplicantProcessProxy callback
This is a second stab at https://crrev.com/c/1377576 (reverted
https://crrev.com/c/1387993). Important change: use a new weak pointer
factory with a longer lifetime, since the original factory is
invalidated in Stop(). The supplicant callbacks (unlike many of our
others) need to survive beyond Stop()/(Re)Start().

Original description still applies, + a new (manual) TEST:

This proxy object maintains its lifetime with the WiFi object, but it
also posts tasks to the event loop using this Closure. This means
there's a possibility of queueing tasks to the event loop that will run
after we are destroyed.

Instead of an unmanaged Unretained(), give the callback a weak pointer,
so the run loop can decide whether or not to run the callback based on
the weak pointer's validity.

This should avoid some nullptr dereferences we've seen, where
OnSupplicantAppear() is called just after its WiFi object was destroyed.

BUG=chromium:698457
TEST=unit tests
TEST=manual test - add/remove Wifi devices (e.g., with 'iw')
TEST=`restart wpasupplicant`; shill re-establishes Wifi services

Change-Id: Ibb7c2ecc25379bf5ed0a96107e9f2ec0e2ab85f6
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1389416
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>