shill: wifi: fix other callback lifetimes across Stop()/Restart()
The Mac80211Monitor and DebugScopeChanged callbacks should be available
even after a device Restart, but today, they use the weak pointer
factory that gets invalidated during Stop().

For OnWiFiDebugScopeChanged(), this callback is safe as long as the
process proxy object is valid, which is the same lifetime as the WiFi
object. So there's no need to invalidate it in Stop().

For Mac80211Monitor, it's currently safe enough to keep its lifetime
(the real work under StartScanTimer() is eventually gated by
'enabled()'), but it's much clearer if the callback itself is gated
appropriately.

Note that this uncovers a bad unit test:
ConnectToServiceWithoutRecentIssues was ensuring that we never touched
the debug level at all (rather than checking that we don't increase it).
This is violated in the test TearDown(), where we reset levels to
'-wifi'. This is a desireable behavior, because as long as we still can
talk to wpa_supplicant (even if we're disconnected), we should be
resetting its debug level. Convert this assertion to just checking for
whether we think we're debugging the connection.

BUG=chromium:698457
TEST=unit tests
TEST=watch scan behavior while connected or disconnected, and restarting
     wpasupplicant

Change-Id: Ibca1acde34e56b9608557a759f71f5c4c04351d6
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1389417
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>