usb_bouncer: Move lockscreen state logic out of userdb initializaiton.
This adds a user_db_read_only flag to EntryManager that allows the
rules.conf to be generated from the user device whitelist database
instead of it falling back on the global database which only includes
currently connected devices.

Unit tests were added to test for the case when the lock screen is
shown.

BUG=chromium:919903
TEST=FEATURES=test emerge-${BOARD} usb_bouncer

Change-Id: I76edc2b5359da41a4453ed0c82b2aabe5d957e12
Reviewed-on: https://chromium-review.googlesource.com/1391439
Commit-Ready: Allen Webb <allenwebb@google.com>
Tested-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>