cros-disks: Clamp time offset option for vfat to +-12 hours.
The fat filesystem on kernels <4.5 refuse to mount partitions when the
time_offset mount option if outside the +-12 hours range. A number of
countries are in the UTC+13 and UTC+14 time zones. Users in these
countries experience mount failures when attempting to access devices
formatted as FAT32.

With this change, users in these time zones will see incorrect time
stamps for files on FAT32 drives. However, any other solution is far
more complicated and the intention for this patch is to be mergable into
M71/M72 if needed. Further, this patch will be reverted once the
relavent kernel patch is back ported to all current kernels.

BUG=chromium:916643
TEST=Unit tests + Deploy to chell, change time zone to Auckland, insert
     FAT32 formatted USB drive.

Change-Id: I45f32b6723eeb30a1ae453f3d47c7541a85542d2
Reviewed-on: https://chromium-review.googlesource.com/1392574
Commit-Ready: Anand Mistry <amistry@chromium.org>
Tested-by: Anand Mistry <amistry@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>