shill: wifi: rename weak_ptr_factory_
Common weak pointer usage aligns an object's weak pointer lifetime with
the lifetime of the object itself, but the usage in WiFi is a bit
non-traditional, instead following the lifetime of device
Start()/Stop(). This has helped to introduce subtle bugs that only
appear after we disable/re-enable the device (e.g., in the network UI).

I plan to introduce a new weak pointer factory that follows the WiFi
object lifetime (and fixes several bugs), so let's rename the existing
one. I do this in a separate CL to separate the functional changes from
this noisy rename/refactor.

BUG=chromium:698457
TEST=unit tests

Change-Id: Ice5f4fac5132fc180779ea6bd086eeffad9fcfa0
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1393908
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>