cryptohome: remove UnmountForUser
There's a comment on this method questioning whether it should
even exist, and it has no external users. It's not even exposed to
Chrome via system_api. Thus, it's dead, and we should remove it.
If we want this functionality later, we should reimplement it in a
new-style Ex method.

BUG=chromium:811551
TEST=unit tests, pre-cq

Change-Id: Ie16c1a4d1c07545d15acfe7375e6e7c2b9c2ea43
Reviewed-on: https://chromium-review.googlesource.com/1395045
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>