tpm_manager: add owner and endorsement passwords to the ownership taken signal
Add the restriction to the signal: only (process run by) root can
receive the signal.

This is a pre-work of tpm status refactoring/optimization in
cryptohome and attestation. By including the passwords in the signal,
client daemons don't need to actively poll the passwords.

BUG=None
TEST=unit tests;
     Run the following command as root and chronos respectively and
     check that only root can receive the signal:
     $ dbus-monitor --system type=signal \
           sender=org.chromium.TpmManager \
           interface=org.chromium.TpmOwnership \
           member=SignalOwnershipTaken
     (unit tests on the way...)

Change-Id: I98fb5d1a4bf24b38189378dd7d0755a7ce485075
Reviewed-on: https://chromium-review.googlesource.com/1429539
Commit-Ready: Wei-Cheng Xiao <garryxiao@chromium.org>
Tested-by: Wei-Cheng Xiao <garryxiao@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>