oobe_config: Move OobeCompletedFile constant to rollback_constants
This change moves the OobeCompletedFile constant to
rollback_constants so that it can be resued elsewhere. Additionally,
since the constant is always used in the context of a FilePath, this
changes the constant to existing as a base::FilePath rather than a
string.
This CL changes no logic.

BUG=chromium:881341
TEST=none
Change-Id: If02dbb898a1d31dfd88a797d47e8dc322cdbc0b8
Reviewed-on: https://chromium-review.googlesource.com/1405832
Commit-Ready: Bailey Berro <baileyberro@chromium.org>
Tested-by: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Zentaro Kavanagh <zentaro@chromium.org>