vm_tools: Cicerone Fuzz Test
Fuzz test for the Cicerone's ContainerListenerImpl. This is a boundary
between the untrusted code in the container and the trusted VM code, so
it's important that it be secure.

It's not perfect; the fuzz test can't actually start another VM or
Container even though it can exercise some of the parts. But it hits a
lot of the code, and it's better than the nothing we have now.

(ebuild change in crrev.com/c/1407457)

BUG=chromium:904028
TEST=Ran fuzz test

Change-Id: I55e0e8c0890b597d139dbdfcbd7e71b27fd990de
Reviewed-on: https://chromium-review.googlesource.com/1407456
Commit-Ready: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>