init: chromeos_startup: drop cryptohome scanning for now
Since cryptohome paths can get to be quite large, and they're still
encrypted at this point, lets drop the file attributes scanning of
those paths.  We'll need to revisit anyways with crbug.com/919486,
but that'll run as part of cryptohome and not in the normal boot
critical path.

BUG=chromium:915846
TEST=precq passes

Change-Id: I5c104997132603fb139a220edc9e720d619f9be6
Reviewed-on: https://chromium-review.googlesource.com/1405649
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>