shill: cellular: Set IPConfig::Properties::method based on IPConfig::method
The networking UI code relies on the IPConfig::Properties::method
string (through DBus) in order to determine various pieces of
information about a device's IP addresses. In cellular, this string
was not being set to match the IPConfig::method value provided by
modemmanager. This change fixes that, such that the networking UI
behaves as it does in WiFi when a device has both IPv4 and v6
addresses.

BUG=chromium:919665
TEST=Cellular IP address information displayed by the UI is now
     consistent with WiFi when the device has both IPv4 and v6
     addresses (i.e. default IP Address field shows the v4 address,
     and the network tab displays v4 address under "IP address"
     and v6 address under "IPv6 address").

Change-Id: I282e185ca71ca0777507f1bf260665af15ff21b8
Reviewed-on: https://chromium-review.googlesource.com/1405370
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>