shill: Use EthernetProvider
Calling the Manager dbus method ConfigureServiceForProfile on an
ethernet service without EAP authentication would result in an error.
This caused problems for users who wanted to save static IP configs
or DNS nameservers. We fix this in this patch by transition Manager
to use the EthernetProvider class.

Prior to this change, the Ethernet device acted as a provider for
Ethernet. However, this would not allow Chrome to configure a service
before the device came up. EthernetProvider is now in charge of creating
EthernetService objects so that Chrome can push down static IP configs
before Ethernet is plugged in. It does this by first creating a generic
EthernetService object without an associated device (similar to WiFi).
This service, created with storage ID "ethernet_any", is used as a
template for all other EthernetService objects. When an EthernetService
with an associated device is created, it loads the static IP config
from the generic profile.

As of now, we haven't found a way to support multiple Ethernet devices.
We simply pick the first one according to shill's service sorting
algorithm when deciding which one to configure.

BUG=chromium:916041
TEST=1) unit tests
     2) dbus-send --print-reply --system --dest=org.chromium.flimflam /
          org.chromium.flimflam.Manager.ConfigureServiceForProfile
          objpath:/profile/default
          dict:string:variant:Type,string:ethernet,StaticIP.Address,string:100.90.25.228

        This returns an object path pointing to the ethernet service, as
        expected. `ifconfig` confirms that the correct IP was set. Running
        this before/after plugging in Ethernet works, and the changes
        are synced across the two profiles.

Change-Id: If404252bffde8757454b53282bcdd20b575a5cfc
Reviewed-on: https://chromium-review.googlesource.com/1407822
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>