vm_tools: Speed up ServiceTestingHelper destructor
The fuzz test in
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1407456
spends approximatly 99% of its run time in the destructor for
ServiceTestingHelper. This CL changes the destructor to shutdown the
tremplin server before waiting for the
ServiceTestingHelper::DestroyService to complete. This reduces the run
time of the method by around 30%, from ~3 seconds to ~2 seconds.

BUG=chromium:904028
TEST=Ran unit tests, ran fuzz test while timing this method

Change-Id: I809c1c46f58ad0bcac5e2169efd3281d738fdda8
Reviewed-on: https://chromium-review.googlesource.com/1409669
Commit-Ready: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>