camera: Update platform2 related paths.
This reverts commit 9fe2dec9926e921379764a798a1884e509df67c3.
and then add more fixes.

The repository platform/arc-camera is migrated into platform2/camera,
but some paths in the code was not updated and kept as they were relative
to old path.

Also, some of the paths contain "../../platform2/common-mk" to refer it
from "platform/arc-camera". It should work from the new location, but
"../platform2" is redundant now.
Specifically, it could cause build issue if out-of-tree build
does not work due to crbug.com/916471.
The actual problematic scenario we hit was written in crbug.com/922274#4.
Thus, the redundant "../platform2" is removed as the fix for it,
which is the diff from the reverted CL.

BUG=chromium:920521, chromium:803783
TEST=Ran try.
CQ-DEPEND=CL:1414530, CL:1414511

Change-Id: I8d8a7d9398b8e3aacf997063030ed76d85327c38
Reviewed-on: https://chromium-review.googlesource.com/1414255
Commit-Ready: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>